Misc. fixes after review of ckeditor5-find-and-replace.
