+++ .gitignore
@@ -45,8 +45,8 @@ packages/ckeditor5-editor-inline/src/**/*.js
 packages/ckeditor5-engine/src/**/*.js
 packages/ckeditor5-enter/src/**/*.js
 packages/ckeditor5-essentials/src/**/*.js
-packages/ckeditor5-font/src/**/*.js
 packages/ckeditor5-find-and-replace/src/**/*.js
+packages/ckeditor5-font/src/**/*.js
 packages/ckeditor5-heading/src/**/*.js
 packages/ckeditor5-highlight/src/**/*.js
 packages/ckeditor5-horizontal-line/src/**/*.js
